feat(cmds/status): combine deliver.go,prepare.go, and acceptorder.go into a single command
